separate ActivityLevel from AppUser
