+++ backend/src/main/java/com/github/ingbeck/backend/model/AppUser.java
@@ -42,20 +42,3 @@ public class AppUser {
 
 }
 
-enum ActivityLevel
-{
-    ATHLETE(1.725),
-    PUMPER(1.55),
-    PEDESTRIAN(1.375),
-    COUCHPOTATO(1.2);
-
-    private double level;
-
-    ActivityLevel(double level) {
-        this.level = level;
-    }
-
-    public double getLevel() {
-        return level;
-    }
-}
